+++ CHANGELOG.md
@@ -1,7 +1,10 @@
 # CPSC436 Project
 
 All changes will be logged here
-
+##0.12.0 [2020-06-26]
+### Add frontend UI for login + register page and change APP Router accordingly
+- add front UI login register
+- change APP router
 ##0.11.0 [2020-06-25]
 ### Added A number of redux functionality to post page including:
 - add post to post list